평점관련 이슈가 상당히 많음. 1) test위해 seed 설치 2) rank floatfield로 seed도 username쪽에 문제가 있어서 이를 해결하려고하는중

+++ server/movies/models.py
@@ -10,14 +10,14 @@ class Movie(models.Model):
     pubDate = models.CharField(max_length=256)
     director = models.CharField(max_length=256)
     actor = models.CharField(max_length=256)
-    userRating = models.IntegerField()
+    userRating = models.FloatField()
 
 
 class Review(models.Model):
     user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
     movie = models.ForeignKey(Movie, on_delete=models.CASCADE)
     title = models.CharField(max_length=256)
-    rank = models.IntegerField()
+    rank = models.FloatField()
     content = models.TextField()
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
